requirements.txt
backcall==0.1.0
decorator==4.3.2
Django==2.1.7
django-model-utils==3.1.2
djangorestframework==3.9.1
fancycompleter==0.8
jedi==0.13.3
mysqlclient==1.4.2.post1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
psycopg2==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
pytz==2018.9
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
wmctrl==0.3
gunicorn==19.9.0
whitenoise==4.1.2
django-storages==1.7.1
dropbox==9.3.0
django-debug-toolbar==1.11